Life and Metal Day 1
Added various files, this includes the current scene set up. Using mostly images off the internet I created a small map and character, these images will be changed around by me in a few days so that they are original and fit what I need. There is a small collision box around the main character sprite. As of yet there is no code behind this and does not do much.